// web/src/types.ts
export interface PlanetData {
  id: string
  name: string
  massKg: number
  radiusKm: number
  temperatureK: number
  distanceAu: number
  orbitalPeriodDays: number
  surfaceGravity: number
  color: string
  hasRings: boolean
  eccentricity?: number
  inclinationDeg?: number
  meanLongitudeJ2000Deg?: number
  perihelionLongitudeDeg?: number
  ascendingNodeLongitudeDeg?: number
}